requirements.txt
Pillow==7.1.2
reportlab==3.5.42
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
